Even cooler stuff

+++ raytracer/scene.h
@@ -41,7 +41,7 @@ private:
 	Color globalAmbient;
 	double goochB, goochY, goochAlpha, goochBeta;
 	double edges;
-	unsigned int photonFactor;
+	int photonFactor;
 	double photonIntensity;
 	
 	Color calcPhong(Object *obj, Point *hit, Vector *N, Vector *V, unsigned int recursionDepth, double recursionWeight);
@@ -91,7 +91,7 @@ public:
 	void setEdges(double e) { edges = e; }
 	void setMaxRecursionDepth(unsigned int d) { maxRecursionDepth = d; }
 	void setMinRecursionWeight(double w) { minRecursionWeight = w; }
-	void setPhotonFactor(unsigned int f) { photonFactor = f; }
+	void setPhotonFactor(unsigned int f) { photonFactor = (int)f; }
 	void setPhotonIntensity(double i) { photonIntensity = i; }
 	unsigned int getNumObjects() { return objects.size(); }
 	unsigned int getNumLights() { return lights.size(); }
